Add more detailed descriptions in README.md and comments in code. Also rename uart.cpp -> uart.c and remove some duplicate code with macro.
