+++ emulator/clients/clients.mk
@@ -1,7 +1,7 @@
 CLIENT_DIR ?= $(EMULATOR_DIR)/clients
 CLIENT_BUILD_DIR  ?= $(CLIENT_DIR)/build
-CLIENT_C_SRCS   = interrupter.c
-CLIENT_CPP_SRCS = uart.cpp
+CLIENT_C_SRCS   = interrupter.c uart.c
+CLIENT_CPP_SRCS = 
 COMMON_SRCS = common.c
 
 CLIENTS  = $(patsubst %.c,$(CLIENT_BUILD_DIR)/%.elf,$(CLIENT_C_SRCS))
