+++ emulator/clients/common.h
@@ -9,8 +9,36 @@
 #define LOW  (0)
 
 struct PinEvent {
+    /**
+     * Which pin the event occurs on; this number is determined by the macros 
+     * found below.
+     * 
+     */
     uint32_t pin;
+
+    /**
+     * The number of cycles until the event should occur relative to the time
+     * of the last PinEvent in the list of pending events.
+     * 
+     * E.g., if the list already contains three events like this:
+     * 
+     *  { 
+     *      { .pin = MY_PIN, .in_n_cycles = 300, .high_low = HIGH },
+     *      { .pin = MY_PIN, .in_n_cycles = 500, .high_low = LOW  },
+     *      { .pin = MY_PIN, .in_n_cycles = 250, .high_low = HIGH },
+     *  };
+     * 
+     * when the emulator receives the list it will start counting. When it hits
+     * 300 it will trigger the first event and restart the counter. When it hits
+     * 500 it will trigger the second event and restart the counter. And so on.
+     * 
+     */
     uint32_t in_n_cycles;
+
+    /**
+     * Whether to set the pin HIGH (1) or LOW (0).
+     * 
+     */
     bool high_low;
 };
